pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mkdocs-media-gallery-plugin"
version = "1.0.0.dev30"
description = "MkDocs plugin providing media galleries with shortcodes"
authors = [{ name = "APinchofDill" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "mkdocs>=1.5",
  "jinja2>=3.0",
  "pyyaml>=6.0",
  "mkdocs-glightbox>=0.4.0"
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
keywords = ["mkdocs", "image", "gallery"]

[project.urls]
Homepage = "https://github.com/APinchofDill/mkdocs-media-gallery-plugin"

[project.entry-points."mkdocs.plugins"]
media-gallery = "mkdocs_media_gallery.plugin:MediaGalleryPlugin"

[tool.setuptools.packages.find]
include = [
  "mkdocs_media_gallery",
  "mkdocs_media_gallery.*"
]

[tool.setuptools.package-data]
mkdocs_media_gallery = [
  "assets/*",
  "templates/*.html"
]
